requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

confluent-kafka
deap
easydict
filterpy
lap
loguru
matplotlib
moviepy
numpy
nvidia-modelopt[all]
onnxruntime_gpu
pillow
pybind11
pybind11-stubgen
pyyaml
scikit-image
scikit_learn
scipy
shapely
tabulate
tensorboard
torch==2.7.1
torch_tensorrt==2.7.0
torchinfo
torchvision
tqdm